rust: move `file_operations::FileOperations` into `file::Operations`
This is to avoid repeating the "file" prefix and for consistency with
other modules that also have operations like `power::Operations`.

This is a pure refactor, no functional changes are intended.
