+++ drivers/char/hw_random/bcm2835_rng_rust.rs
@@ -3,8 +3,8 @@
 //! Broadcom BCM2835 Random Number Generator support.
 
 use kernel::{
-    device, file::File, file_operations::FileOperations, io_buffer::IoBufferWriter, miscdev,
-    module_platform_driver, of, platform, prelude::*, sync::Ref,
+    device, file, file::File, io_buffer::IoBufferWriter, miscdev, module_platform_driver, of,
+    platform, prelude::*, sync::Ref,
 };
 
 module_platform_driver! {
@@ -17,7 +17,7 @@ module_platform_driver! {
 
 struct RngDevice;
 
-impl FileOperations for RngDevice {
+impl file::Operations for RngDevice {
     kernel::declare_file_operations!(read);
 
     fn open(_open_data: &(), _file: &File) -> Result {
