+++ rust/kernel/chrdev.rs
@@ -16,7 +16,7 @@ use core::pin::Pin;
 use crate::bindings;
 use crate::c_types;
 use crate::error::{Error, Result};
-use crate::file_operations;
+use crate::file;
 use crate::str::CStr;
 
 /// Character device.
@@ -134,9 +134,7 @@ impl<const N: usize> Registration<{ N }> {
     /// Registers a character device.
     ///
     /// You may call this once per device type, up to `N` times.
-    pub fn register<T: file_operations::FileOperations<OpenData = ()>>(
-        self: Pin<&mut Self>,
-    ) -> Result {
+    pub fn register<T: file::Operations<OpenData = ()>>(self: Pin<&mut Self>) -> Result {
         // SAFETY: We must ensure that we never move out of `this`.
         let this = unsafe { self.get_unchecked_mut() };
         if this.inner.is_none() {
@@ -170,7 +168,7 @@ impl<const N: usize> Registration<{ N }> {
 
         // SAFETY: The adapter doesn't retrieve any state yet, so it's compatible with any
         // registration.
-        let fops = unsafe { file_operations::FileOperationsVtable::<Self, T>::build() };
+        let fops = unsafe { file::OperationsVtable::<Self, T>::build() };
         let mut cdev = Cdev::alloc(fops, this.this_module)?;
         cdev.add(inner.dev + inner.used as bindings::dev_t, 1)?;
         inner.cdevs[inner.used].replace(cdev);
@@ -179,7 +177,7 @@ impl<const N: usize> Registration<{ N }> {
     }
 }
 
-impl<const N: usize> file_operations::FileOpenAdapter<()> for Registration<{ N }> {
+impl<const N: usize> file::OpenAdapter<()> for Registration<{ N }> {
     unsafe fn convert(_inode: *mut bindings::inode, _file: *mut bindings::file) -> *const () {
         // TODO: Update the SAFETY comment on the call to `FileOperationsVTable::build` above once
         // this is updated to retrieve state.
